requirements.txt
# Monocular Depth Estimation – Dependencies
# Core
numpy>=1.24
Pillow>=9.0
scipy>=1.10
opencv-python>=4.7
torch>=2.0
torchvision>=0.15
transformers>=4.35

# Development
tqdm
matplotlib
